% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-infection-network.R
\name{animate_infection_network}
\alias{animate_infection_network}
\title{Creates interactive html file with d3 network animation}
\usage{
animate_infection_network(first_infection_list, file, year = NULL,
  detach = FALSE)
}
\arguments{
\item{first_infection_list}{Infection list outputted by \code{first_infection_list}.}

\item{file}{Full file path where html is to be saved to.}

\item{year}{Year the outbreak relates to for title plotting. Default = NULL}

\item{detach}{Boolean that detaches the packages loaded within the library.
Deafult = FALSE, i.e. packages remain in Global Environment.}
}
\description{
\code{animate_infection_network} takes output of \code{first_infection_list} and
produces a network animation of the outbreak using d3. Output is saved to file
as an html page
}
\details{
N.B. Function
}

