% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-simulated-plot.R
\name{bootstrap_simulated_plot}
\alias{bootstrap_simulated_plot}
\title{Bootstrap and plot simulated epidemic}
\usage{
bootstrap_simulated_plot(R0 = 1.8, I = 3, first_infection_list,
  outbreak.dataset, replicates = 2000, sampling = FALSE,
  lower.quantile = 0.25, upper.quantile = 0.75, title = NULL,
  alpha = 0.2, size = 1, include.line = TRUE, include.observed = FALSE,
  exponential = FALSE)
}
\arguments{
\item{R0}{Reproductive number. Default = 1.8}

\item{I}{Number of initial seed infections. Default = 3}

\item{first_infection_list}{Infection list outputted by \code{first_infection_list}}

\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{replicates}{Numerical describing number of bootstrap replicates. Default = 2000}

\item{sampling}{Boolean determining if recovery and generation times should be sampled
from the observed or drawn fro a poisson with mean equal to mean of the observed.
Default = FALSE (poisson draws used)}

\item{lower.quantile}{Numeric between 0 and 0.5 describing the lower quantile for
each trace. Default = 0.25}

\item{upper.quantile}{Numeric between 0.5 and 1 describing the lower quantile for
each trace. Default = 0.75}

\item{title}{Plot title. If NULL then no title will be added. Defult = NULL}

\item{alpha}{Translucency of quantiles. Default = 0.2}

\item{size}{Line width. Default = 1}

\item{include.line}{Boolean describing whether to plot with the line or just the
quantiles. Default = TRUE}

\item{include.observed}{Boolean describing whether to include the observed epidemic
alongisde the simulated epidemic. If this is true, N will be calculated from outbreak.dataset}

\item{exponential}{Boolean determining if infection is exponential or not within
the simulation. If FALSE (Default) then the number of secondary infections from an individual is
takes into account S/N}
}
\description{
\code{bootstrap_simulated_plot} takes output of \code{first_infection_list} and
\code{outbreak_dataset_read} and uses \code{discrete_SIR_simulator} to simulate
a number of SIR epidemics, which are then summarised and plotted. Returns ggplot
object for plotting.
}
