% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined-plot.R
\name{combined_plot}
\alias{combined_plot}
\title{Plot multiplot of network at 24 hour intervals with the epidemic timeseries}
\usage{
combined_plot(first_infection_list)
}
\arguments{
\item{first_infection_list}{Infection list outputted by \code{first_infection_list}}
}
\description{
\code{combined_plot} takes output of \code{first_infection_list} and
plots multiplot of network at 24 hour intervals with the epidemic timeseries
}
