% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-timeseries-plot.R
\name{daily_timeseries_plot}
\alias{daily_timeseries_plot}
\title{Plot 5x1 multiplot of network at 24 hour intervals}
\usage{
daily_timeseries_plot(first_infection_list)
}
\arguments{
\item{first_infection_list}{Infection list outputted by \code{first_infection_list}}
}
\description{
\code{daily_timeseries_plot} takes output of \code{first_infection_list} and
plots 5x1 plot of infection network at 24 hour intervals
}

