% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-SIR-simulator.R
\name{discrete_SIR_simulator}
\alias{discrete_SIR_simulator}
\title{Discrete in time SIR simulator}
\usage{
discrete_SIR_simulator(R0 = 1.8, N = NULL, I = 3, seed.hour = NULL,
  first_infection_list, outbreak.dataset, sampling = FALSE,
  exponential = FALSE)
}
\arguments{
\item{R0}{Reproductive number. Default = 1.8}

\item{N}{Total population size. If NULL (default) then the total population
is the same as the provided datasets.}

\item{I}{Number of initial seed infections. Default = 3}

\item{seed.hour}{Hour at which seeding of epidemic begins. Default = 9}

\item{first_infection_list}{Infection list outputted by \code{first_infection_list}}

\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{sampling}{Boolean determining if recovery and generation times should be sampled
from the observed or drawn fro a poisson with mean equal to mean of the observed.
Default = FALSE (poisson draws used)}

\item{exponential}{Boolean determining if infection is exponential or not. If
FALSE (Default) then the number of secondary infections from an individual is
takes into account S/N}
}
\description{
\code{discrete_SIR_simulator} takes output of \code{first_infection_list} and
\code{outbreak_dataset_read} and simulates a discrete time SIR epidemic time
series using a user-provided R0, population size (N), number of initial seed
infections (I) and time of seeding (seed.hour). The simulation uses a poisson
distribrution to describe the number of secondary cases, the generation times
and recovery times. The mean for the secondary cases is equal to R0, and the
mean for the generation and recovery times is calculated from the user-provided
first_infection_list and outbreak.dataset.
}
