% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic-timeseries-plot.R
\name{epidemic_timeseries_plot}
\alias{epidemic_timeseries_plot}
\title{Plot epidemic time series}
\usage{
epidemic_timeseries_plot(first_infection_list, N, title = NULL)
}
\arguments{
\item{first_infection_list}{Infection list outputted by \code{first_infection_list}}

\item{N}{Total population size.}

\item{title}{Plot title. If NULL then no title will be added. Defult = NULL}
}
\description{
\code{epidemic_timeseries_plot} takes output of \code{first_infection_list} and
plots SIR epidemic time series. Returns ggplot object for plotting.
}

