% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-infection-list.R
\name{first_infection_list}
\alias{first_infection_list}
\title{Format read outbreak dataset into first infections}
\usage{
first_infection_list(outbreak.dataset)
}
\arguments{
\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}
}
\description{
\code{first_infection_list} takes output of \code{outbreak_dataset_read} and
splits into linelist and contacts for further use, removing reinfections.
}
