% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation-time-daily-plot.R
\name{generation_time_daily_plot}
\alias{generation_time_daily_plot}
\title{Plot Daily Generation Time}
\usage{
generation_time_daily_plot(outbreak.dataset, title = NULL, alpha = NA,
  jitter.width = 0.2, half.day = FALSE)
}
\arguments{
\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{title}{Plot title. Default = NULL}

\item{alpha}{Numeric for plot point transparency. Default = NA, i.e. no transparency}

\item{jitter.width}{Numeric for plot point jitter width. Default = 0.2}

\item{half.day}{Boolean detailing whether to break days into AM and PM. Default = FALSE}
}
\description{
\code{generation_time_daily_plot} takes output of \code{outbreak_dataset_read} and
plots dotplots for each day showing he generation times
}
