% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring-distribution-plot.R
\name{offspring_distribution_plot}
\alias{offspring_distribution_plot}
\title{Plot offspring distribution}
\usage{
offspring_distribution_plot(outbreak.dataset, include.reinfectons = TRUE,
  title = NULL)
}
\arguments{
\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{include.reinfectons}{Boolean detailing whether to include reinfections. Default = TRUE}

\item{title}{Plot title. Default = NULL}
}
\description{
\code{offspring_distribution_plot} takes output of \code{outbreak_dataset_read} and
plots offspring distribution. Returns ggplot object for plotting.
}
