% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramater-boxplots-plot.R
\name{paramater_boxplots_plot}
\alias{paramater_boxplots_plot}
\title{Plot 4x1 boxplots of epidemiological parameters}
\usage{
paramater_boxplots_plot(outbreak.dataset, title = NULL, alpha = NA,
  jitter.width = 0.2)
}
\arguments{
\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{title}{Plot title. Default = NULL}

\item{alpha}{Numeric for plot point transparency. Default = NA, i.e. no transparency}

\item{jitter.width}{Numeric for plot point jitter width. Default = 0.2}
}
\description{
\code{paramater_boxplots_plot} takes output of \code{outbreak_dataset_read} and
plots boxplot of epidemiological parameter observations. Returns ggplot
object for plotting.
}

