% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic-timeseries-plot.R
\name{epidemic_timeseries_plot}
\alias{epidemic_timeseries_plot}
\title{Plot epidemic time series}
\usage{
epidemic_timeseries_plot(outbreak.dataset, title = NULL)
}
\arguments{
\item{outbreak.dataset}{Outbreak Dataset}

\item{title}{Plot title. If NULL then no title will be added. Defult = NULL}
}
\description{
\code{epidemic_timeseries_plot} takes output of \code{outbreak_dataset_read} and
plots SIR epidemic time series. Returns ggplot object for plotting.
}

